requirements.txt
# Core dependencies
openai==1.54.3
anthropic==0.34.2
google-generativeai==0.8.2
python-dotenv==1.0.1

# Explicitly pin httpx to compatible version
httpx==0.27.2

# Data processing
pandas==2.2.3
numpy==2.1.2
scipy==1.14.1

# Statistical analysis
statsmodels==0.14.3
scikit-learn==1.5.2

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Jupyter
jupyter==1.1.1
ipykernel==6.29.5
nbconvert==7.16.4

# Code quality
black==24.10.0
flake8==7.1.1
pytest==8.3.3

# Utilities
tqdm==4.66.5
requests==2.32.3